Totally finish log and Code note.

+++ src/main/kotlin/pw/lunzi/privatecloudstorage/ItemController.kt
@@ -9,7 +9,6 @@ import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.web.bind.annotation.*
 import java.util.*
 import javax.servlet.http.HttpServletRequest
-import kotlin.math.log
 
 @RestController
 class ItemController(private val fileItemRepo: FileItemRepository) {
@@ -42,7 +41,7 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
         //检查目录是否存在
         val superItem: FileItem? = Utils.getSuperItem(path, fileItemRepo)
         if (superItem == null) {
-            itemEditLog.warn("Get index from path \"$path\" failed. Path is invalid")
+            itemEditLog.warn("User \"${if (user != null) user.username else "Guest"}\" Get index from path \"$path\" failed. Path is invalid")
             return ResponseEntity(ReplyMsg(false, "item is invalid"), HttpStatus.NOT_FOUND)
         }
 
@@ -76,19 +75,32 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
         return ResponseEntity(dataList, HttpStatus.OK)
     }
 
+    /**
+     * 重命名Item
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @PutMapping("/api/item/**/name")
     fun rename(@AuthenticationPrincipal user: UserDetails, @RequestBody msg: RenameMsg, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中的**部分，即完整路径
         val completePath = Utils.getPath(Utils.extractPathFromPattern(request))
         val path = Utils.getPath(completePath)
         val name = Utils.getName(completePath)
+        itemEditLog.info("User \"${user.username}\" try to rename item \"$path$name\" to $path${msg.newName}.")
 
-        val fileItem: FileItem = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
-                ?: return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        //获取对应路径的Item
+        val fileItem: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
+        if (fileItem == null) {
+            itemEditLog.warn("User \"${user.username}\" Rename item \"$path$name\" failed. Path is invalid")
+            return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        }
 
-        if (fileItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN")))
+        //若未登录或未拥有修改Item的权限
+        if (fileItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User \"${user.username}\" Rename item \"$path$name\" failed. Permission denied")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
+        }
 
+        //若重命名的为文件夹，递归更新子item路径
         if (fileItem.isDictionary) {
             val oldPath: CharSequence = path + name
             fileItemRepo.findByOwnerName(user.username).forEach {
@@ -105,23 +117,36 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
         fileItem.lastModified = Date()
         fileItemRepo.save(fileItem)
 
+        itemEditLog.info("User \"${user.username}\" rename item \"$path$name\" to $path${msg.newName} success.")
         return ResponseEntity(ReplyMsg(true, "Rename $path$name to $path${msg.newName} success"), HttpStatus.OK)
     }
 
+    /**
+     * 删除Item
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @DeleteMapping("/api/item/**")
     fun delete(@AuthenticationPrincipal user: UserDetails, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中的**部分路径
         val completePath = Utils.extractPathFromPattern(request)
         val path = Utils.getPath(completePath)
         val name = Utils.getName(completePath)
+        itemEditLog.info("User \"${user.username}\" try to delete item \"$path$name\".")
 
-        val fileItem: FileItem = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
-                ?: return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        //检查Item是否存在
+        val fileItem: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
+        if (fileItem == null) {
+            itemEditLog.warn("User \"${user.username}\" delete item \"$path$name\" failed. Path is invalid")
+            return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        }
 
+        //检查用户是否拥有权限
         if (fileItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User \"${user.username}\" delete item \"$path$name\" failed. Permission deneid")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
         }
 
+        //若Item为文件夹，递归删除其子Item
         if (fileItem.isDictionary) {
             fileItemRepo.delete(fileItem)
             var count = 1
@@ -133,32 +158,45 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
                     if (fileItemRepo.findByRealPath(it.realPath).isEmpty() && !it.isDictionary) it.deleteFile()
                 }
             }
-            return ResponseEntity(ReplyMsg(true, "Delete folder ${fileItem.virtualPath}${fileItem.virtualName}/ total $count items success"), HttpStatus.OK)
+            itemEditLog.info("User \"${user.username}\" delete item \"$path$name\" success. Total $count items")
+            return ResponseEntity(ReplyMsg(true, "Delete folder ${path}${name}/ total $count items success"), HttpStatus.OK)
         } else {
             fileItemRepo.delete(fileItem)
             Utils.updateSize(fileItem, -1 * fileItem.size, fileItemRepo)
             if (fileItemRepo.findByRealPath(fileItem.realPath).isEmpty()) fileItem.deleteFile()
+            itemEditLog.info("User \"${user.username}\" delete item \"$path$name\" success.")
             return ResponseEntity(ReplyMsg(true, "Delete ${fileItem.virtualPath}${fileItem.virtualName} success"), HttpStatus.OK)
         }
     }
 
+    /**
+     * 移动Item
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @PutMapping("/api/item/**/path")
     fun move(@AuthenticationPrincipal user: UserDetails, @RequestBody msg: MoveMsg, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中的**部分路径
         val completePath = Utils.getPath(Utils.extractPathFromPattern(request))
         val path = Utils.getPath(completePath)
         val name = Utils.getName(completePath)
 
+        itemEditLog.info("User \"${user.username}\" try to move item \"$path$name\" to \"${msg.newPath}$name\".")
+
         //Check origin file is legal
-        val fileItem: FileItem = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
-                ?: return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        val fileItem: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
+        if (fileItem == null) {
+            itemEditLog.warn("User \"${user.username}\" move item \"$path$name\" to \"${msg.newPath}$name\" failed. Origin path is invalid.")
+            return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        }
 
         //Check new path is legal
         if (Utils.getSuperItem(msg.newPath, fileItemRepo) == null) {
+            itemEditLog.warn("User \"${user.username}\" move item \"$path$name\" to \"${msg.newPath}$name\" failed. New path is invalid.")
             return ResponseEntity(ReplyMsg(false, "Sorry. New path is invalid"), HttpStatus.BAD_REQUEST)
         }
 
         if (fileItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User \"${user.username}\" move item \"$path$name\" to \"${msg.newPath}$name\" failed. Permission denied.")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
         }
 
@@ -185,38 +223,55 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
         Utils.updateSize(fileItem, fileItem.size, fileItemRepo)
         fileItemRepo.save(fileItem)
 
+        itemEditLog.info("User ${user.username} move $path$name to ${msg.newPath}$name total $count ${if (count == 1) "item" else "items"} success")
         return ResponseEntity(ReplyMsg(true, "Move $path$name to ${msg.newPath}$name total $count ${if (count == 1) "item" else "items"} success"), HttpStatus.OK)
     }
 
+    /**
+     * 更改Item权限（private/public）
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @PutMapping("/api/item/**/access")
     fun changeAccess(@AuthenticationPrincipal user: UserDetails, @RequestBody msg: ChangeAccessMsg, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中的**部分路径
         val completePath = Utils.getPath(Utils.extractPathFromPattern(request))
         val path = Utils.getPath(completePath)
         val name = Utils.getName(completePath)
 
-        //Check origin file is legal
-        val fileItem: FileItem = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
-                ?: return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        itemEditLog.info("User ${user.username} try to change \"$path$name\" access to \"${if (msg.isPublic) "public" else " private"}\".")
 
+
+        //Check origin file is legal
+        val fileItem: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
+        if (fileItem == null) {
+            itemEditLog.warn("User \"${user.username}\" change  \"$path$name\" access to \"${if (msg.isPublic) "public" else " private"}\" failed. Item is invalid.")
+            return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        }
+
+        //Check item belong
         if (fileItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User \"${user.username}\" change  \"$path$name\" access to \"${if (msg.isPublic) "public" else " private"}\" failed. Permission denied.")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
         }
 
         //Do change
         var count = 1
-        var temp = fileItem
-        while (temp.virtualPath != "/") {
-            temp = Utils.getSuperItem(temp, fileItemRepo) ?: break
-            if (temp.isPublic != msg.isPublic) {
-                fileItemRepo.delete(temp)
-                temp.isPublic = msg.isPublic
-                temp.lastModified = Date()
-                fileItemRepo.save(temp)
-                count++
+        //若将一个Item权限从private更改为public，检查所有父目录，将所有父目录也设为public
+        if (msg.isPublic) {
+            var temp: FileItem = fileItem
+            while (temp.virtualPath != "/") {
+                temp = Utils.getSuperItem(temp, fileItemRepo) ?: break
+                if (temp.isPublic != msg.isPublic) {
+                    fileItemRepo.delete(temp)
+                    temp.isPublic = msg.isPublic
+                    temp.lastModified = Date()
+                    fileItemRepo.save(temp)
+                    count++
+                }
             }
         }
 
+        //若Item为文件夹且参数允许递归更改，递归更改所有子Item的权限
         if (fileItem.isDictionary && msg.allowRecursion) {
             fileItemRepo.findByOwnerName(user.username).forEach {
                 if (it.virtualPath.contains(path + name)) {
@@ -233,31 +288,47 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
         fileItem.lastModified = Date()
         fileItemRepo.save(fileItem)
 
+        itemEditLog.info("User ${user.username} change $path$name total $count ${if (count == 1) "item" else "items"} access to ${if (msg.isPublic) "public" else " private"} success.")
         return ResponseEntity(ReplyMsg(true, "Change $path$name total $count ${if (count == 1) "item" else "items"} access to ${if (msg.isPublic) "public" else " private"} success"), HttpStatus.OK)
     }
 
+    /**
+     * 转存文件
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @PostMapping("/api/items/**")
     fun transfer(@AuthenticationPrincipal user: UserDetails, @RequestBody msg: TransferMsg, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中**部分Path
         val newPath = Utils.extractPathFromPattern(request)
 
-        //Check origin file is legal
-        val fileItem: FileItem = fileItemRepo.findByVirtualPathAndVirtualName(msg.path, msg.name)
-                ?: return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        itemEditLog.info("User ${user.username} try to transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\".")
 
+        //Check origin file is legal
+        val fileItem: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(msg.path, msg.name)
+        if (fileItem == null) {
+            itemEditLog.warn("User ${user.username} transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\" failed. Item is invalid.")
+            return ResponseEntity(ReplyMsg(false, "Sorry. Item is invalid"), HttpStatus.NOT_FOUND)
+        }
         //Check the file is already belong you
-        if (fileItem.ownerName == user.username) return ResponseEntity(ReplyMsg(false, "Item is already belong you"), HttpStatus.ALREADY_REPORTED)
+        if (fileItem.ownerName == user.username) {
+            itemEditLog.warn("User ${user.username} transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\" failed. Item is already belong you.")
+            return ResponseEntity(ReplyMsg(false, "Item is already belong you"), HttpStatus.ALREADY_REPORTED)
+        }
 
         //Check new path is legal
         if (Utils.getSuperItem(newPath, fileItemRepo) == null) {
+            itemEditLog.warn("User ${user.username} transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\" failed. New path is invalid.")
             return ResponseEntity(ReplyMsg(false, "New path is invalid"), HttpStatus.BAD_REQUEST)
         }
 
-        if (!fileItem.isPublic && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN")))
+        if (!fileItem.isPublic && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User ${user.username} transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\" failed. Permission denied.")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
+        }
 
         //Do transfer
         var count = 1
+        //若转存的Item是文件夹，递归转存文件夹下的所有public文件
         if (fileItem.isDictionary) {
             fileItemRepo.findByOwnerName(fileItem.ownerName).forEach {
                 if (it.virtualPath.contains(msg.path + msg.name) && it.isPublic) {
@@ -289,30 +360,44 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
                 Date()
         ))
         Utils.updateSize(fileItem, fileItem.size, fileItemRepo)
-
+        itemEditLog.info("User ${user.username} try to transfer \"${msg.path}${msg.name}\" to \"$newPath${msg.name}\" total $count ${if (count == 1) "item" else "items"} success.")
         return ResponseEntity(ReplyMsg(true, "Transfer ${msg.path}${msg.name} to $newPath${msg.name} total $count ${if (count == 1) "item" else "items"} success"), HttpStatus.OK)
     }
 
+    /**
+     * 创建文件夹
+     */
     @PreAuthorize("hasRole('ROLE_MEMBER')")
     @PostMapping("/api/item/**")
     fun makeDir(@AuthenticationPrincipal user: UserDetails, request: HttpServletRequest): ResponseEntity<ReplyMsg> {
+        //获取URL中**部分Path
         val completePath = Utils.extractPathFromPattern(request)
         val path = Utils.getPath(completePath)
         val name = Utils.getName(completePath)
 
+        itemEditLog.info("User ${user.username} try to create dir \"$completePath\"")
+
         //Check folder is not exist
         val testExist: FileItem? = fileItemRepo.findByVirtualPathAndVirtualName(path, name)
-        if (testExist != null) return ResponseEntity(ReplyMsg(false, "Dictionary is already exist"), HttpStatus.ALREADY_REPORTED)
+        if (testExist != null) {
+            itemEditLog.warn("User ${user.username} create dir \"$completePath\" failed. Dictionary is already exist.")
+            return ResponseEntity(ReplyMsg(false, "Dictionary is already exist"), HttpStatus.ALREADY_REPORTED)
+        }
 
         //Check super path is legal
-        val superItem: FileItem = Utils.getSuperItem(path, fileItemRepo)
-                ?: return ResponseEntity(ReplyMsg(false, "Super path not exist"), HttpStatus.NOT_FOUND)
+        val superItem: FileItem? = Utils.getSuperItem(path, fileItemRepo)
+        if (superItem == null) {
+            itemEditLog.warn("User ${user.username} create dir \"$completePath\" failed. Super path is invalid.")
+            return ResponseEntity(ReplyMsg(false, "Super path not exist"), HttpStatus.NOT_FOUND)
+        }
 
+        //Check permission
         if (superItem.ownerName != user.username && !user.authorities.contains(SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            itemEditLog.warn("User ${user.username} create dir \"$completePath\" failed. Permission denied.")
             return ResponseEntity(ReplyMsg(false, "Permission denied"), HttpStatus.FORBIDDEN)
         }
 
-
+        //Do mkdir
         val newDir = FileItem(
                 user.username,
                 false,
@@ -322,6 +407,8 @@ class ItemController(private val fileItemRepo: FileItemRepository) {
                 virtualName = name
         )
         fileItemRepo.save(newDir)
+
+        itemEditLog.info("User ${user.username} create dir \"$completePath\" success.")
         return ResponseEntity(ReplyMsg(true, "Create dictionary $path$name success"), HttpStatus.OK)
     }
 }
